+++ test/WaterfallRenderContext.test.mjs
@@ -1,7 +1,7 @@
 import { ok, strictEqual } from 'assert';
 import React from 'react';
 import ReactDOMServer from 'react-dom/server.js';
-import WaterfallRenderContext from '../public/WaterfallRenderContext.js';
+import WaterfallRenderContext from '../public/WaterfallRenderContext.mjs';
 import getBundleSize from './getBundleSize.mjs';
 
 export default (tests) => {
@@ -23,7 +23,7 @@ export default (tests) => {
 
   tests.add('`WaterfallRenderContext` bundle size.', async () => {
     const kB = await getBundleSize(
-      new URL('../public/WaterfallRenderContext.js', import.meta.url)
+      new URL('../public/WaterfallRenderContext.mjs', import.meta.url)
     );
 
     ok(kB < 0.5);